Starting over yet again.
So, it's been a while that this repo has existed. The first commit in the log
is from two years ago, and I'm guessing that this repo housed the first
html/css mockups of the interface. A lot has happened since I started this
project, but the most important thing to talk about here would be where this
specific project is going.

Causemap, the web app that allows users to link situations by cause and effect,
has been split into three code projects:

1. **Cartography**: Low-level database and ORM library. It's main offering is a
library of basic models.

2. **Causemap**: Community-centric library that depends and extends on
Cartography. Users, comments, actions, and search ranking based on user
activity (search may at some point be refactored into another library).

3. **Causemap.org**: The web app that relies on the Causemap library. It's what
will be served by causemap.org.

This is the plan for now.